+++ leia-models/src/main/java/com/grookage/leia/models/SchemaConstants.java
@@ -4,6 +4,5 @@ import lombok.experimental.UtilityClass;
 
 @UtilityClass
 public class SchemaConstants {
-    public static final String LATEST_VERSION = "latest";
     public static final String KEY_DELIMITER = ":";
 }