+++ leia-models/src/main/java/com/grookage/leia/models/schema/ingestion/CreateSchemaRequest.java
@@ -40,6 +40,8 @@ import java.util.Set;
 public class CreateSchemaRequest {
 
     @NotBlank
+    private String versionId;
+    @NotBlank
     private String schemaName;
     @NotNull
     private SchemaType schemaType;