+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+
+## [0.0.1-RC22]
+
+- Removed the versionIdGenerator and the `latest` schema fetch
+- Introduced a search request and cleaned up searchable schemas
+- Fixed necessary tests, made the changes to clientSupplier during retrieval as well
+- Added an ignoreCache at a resource level to make sure we control where to fetch the data from
+
 ## [0.0.1-RC21]
 
 - SchemaDefinition: Default version now points to "latest" approved schema so client's don't have to maintain version
-- DefaultMessageProcessor: Added Builder constructor 
+- DefaultMessageProcessor: Added Builder constructor
 
 ## [0.0.1-RC20]